fix(webapp): refactor structure
