--- src/index.js
@@ -1,25 +1,0 @@
-import AccountInfo from './components/AccountInfo'
-import Backdrop from './components/Backdrop'
-import BPJsonGenerator from './components/BPJsonGenerator'
-import CreateAccount from './components/CreateAccount'
-import DropzoneHash from './components/DropzoneHash'
-import InfiniteRegistryTableLoader from './components/InfiniteRegistryTableLoader'
-import InputHash from './components/InputHash/InputHash'
-import IssuanceVerifier from './components/IssuanceVerifier'
-import RicardianContract from './components/RicardianContract'
-import VisualCertificate from './components/VisualCertificate'
-import ArrayTextField from './components/ArrayTextField'
-
-export {
-  AccountInfo,
-  Backdrop,
-  BPJsonGenerator,
-  CreateAccount,
-  DropzoneHash,
-  InfiniteRegistryTableLoader,
-  InputHash,
-  IssuanceVerifier,
-  RicardianContract,
-  VisualCertificate,
-  ArrayTextField
-}
