+++ src/lib/BPJsonGenerator/ImagePreview.js
