+++ src/lib/DropzoneHash/FileComponent.js
