+++ src/lib/DropzoneHash/index.js
