+++ src/lib/DropzoneHash/styles.js
