+++ src/lib/InfiniteRegistryTableLoader/index.js
