+++ src/lib/InfiniteRegistryTableLoader/styles.js
