+++ src/lib/InputHash/index.js
