+++ src/lib/common/index.js
@@ -11,7 +11,7 @@ import VerticalAlignTopIcon from '@material-ui/icons/VerticalAlignTop'
 import Snackbar from '@material-ui/core/Snackbar'
 import Alert from '@material-ui/lab/Alert'
 
-import { parseFile } from '../../utils'
+import { parseFile } from '../utils'
 
 import Styles from './styles'
